requirements.txt
mysql_connector_repackaged==0.3.1
pandas==2.2.2
selenium==4.23.1
setuptools==70.2.0
streamlit==1.36.0
